+++ src/main/java/com/codewise/util/memory/FixedSafeMutableMemory.java
@@ -23,67 +23,67 @@ public class FixedSafeMutableMemory extends AbstractMutableMemory implements Byt
     }
 
     @Override
-    public byte get(int index) {
+    public byte get(long index) {
         checkCapacity(index + Byte.BYTES);
-        return memory[index];
+        return memory[(int) index];
     }
 
     @Override
-    public void put(int index, byte b) {
+    public void put(long index, byte b) {
         ensureCapacity(index + Byte.BYTES);
-        memory[index] = b;
+        memory[(int) index] = b;
     }
 
     @Override
-    public char getChar(int index) {
+    public char getChar(long index) {
         checkCapacity(index + Character.BYTES);
-        return MemoryAccess.getCharSafe(memory, index);
+        return MemoryAccess.getCharSafe(memory, (int) index);
     }
 
     @Override
-    public void putChar(int index, char value) {
+    public void putChar(long index, char value) {
         ensureCapacity(index + Character.BYTES);
-        MemoryAccess.setCharSafe(memory, index, value);
+        MemoryAccess.setCharSafe(memory, (int) index, value);
     }
 
     @Override
-    public short getShort(int index) {
+    public short getShort(long index) {
         checkCapacity(index + Short.BYTES);
-        return MemoryAccess.getShortSafe(memory, index);
+        return MemoryAccess.getShortSafe(memory, (int) index);
     }
 
     @Override
-    public void putShort(int index, short value) {
+    public void putShort(long index, short value) {
         ensureCapacity(index + Short.BYTES);
-        MemoryAccess.setShortSafe(memory, index, value);
+        MemoryAccess.setShortSafe(memory, (int) index, value);
     }
 
     @Override
-    public int getInt(int index) {
+    public int getInt(long index) {
         checkCapacity(index + Integer.BYTES);
-        return MemoryAccess.getIntSafe(memory, index);
+        return MemoryAccess.getIntSafe(memory, (int) index);
     }
 
     @Override
-    public void putInt(int index, int value) {
+    public void putInt(long index, int value) {
         ensureCapacity(index + Integer.BYTES);
-        MemoryAccess.setIntSafe(memory, index, value);
+        MemoryAccess.setIntSafe(memory, (int) index, value);
     }
 
     @Override
-    public long getLong(int index) {
+    public long getLong(long index) {
         checkCapacity(index + Long.BYTES);
-        return MemoryAccess.getLongSafe(memory, index);
+        return MemoryAccess.getLongSafe(memory, (int) index);
     }
 
     @Override
-    public void putLong(int index, long value) {
+    public void putLong(long index, long value) {
         ensureCapacity(index + Long.BYTES);
-        MemoryAccess.setLongSafe(memory, index, value);
+        MemoryAccess.setLongSafe(memory, (int) index, value);
     }
 
     @Override
-    protected byte[] getMemoryPageAsByteArray(int offset) {
+    protected byte[] getMemoryPageAsByteArray(long offset) {
         return memory;
     }
 }
