+++ src/test/java/com/codewise/util/memory/FixedMutableMemoryTestBase.java
@@ -49,7 +49,7 @@ public abstract class FixedMutableMemoryTestBase<M extends AbstractMutableMemory
     @Test(expectedExceptions = BufferUnderflowException.class, enabled = MemoryAccess.RANGE_CHECKS)
     public void shouldUnderflowWhenGetByteArrayFromOutsideOfBuffer() {
         // given
-        byte[] buf = new byte[memory.capacity() + 1];
+        byte[] buf = new byte[(int) memory.capacity() + 1];
 
         // when
         memory.get(Long.BYTES, buf, 0, buf.length);
@@ -59,7 +59,7 @@ public abstract class FixedMutableMemoryTestBase<M extends AbstractMutableMemory
     public void shouldNotUnderflowWhenGetByteArray() {
         // given
         setUpMemory();
-        byte[] buf = new byte[memory.capacity() + 1];
+        byte[] buf = new byte[(int) memory.capacity() + 1];
 
         // when
         memory.get(0, buf, 0, buf.length);
@@ -71,7 +71,7 @@ public abstract class FixedMutableMemoryTestBase<M extends AbstractMutableMemory
     @Test(expectedExceptions = BufferOverflowException.class, enabled = MemoryAccess.RANGE_CHECKS)
     public void shouldOverflowWhenPutByteArray() {
         // given
-        byte[] buf = new byte[memory.capacity() + 1];
+        byte[] buf = new byte[(int) memory.capacity() + 1];
 
         // when
         memory.put(0, buf, 0, buf.length);
